Move ControlSnapshot handler logic into common one
